% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eColiEmbeddingMatrix}
\alias{eColiEmbeddingMatrix}
\title{Raw Per-Protein Embeddings For \emph{Escherichia coli}}
\format{
A data frame with 4,401 rows and 1024 total columns. In the index are
    UniProt identifiers for each protein, and columns are labelled V1, V2,
    ..., V1024 to represent each dimension of the embedding.
}
\source{
UniProt Consortium.
}
\usage{
eColiEmbeddingMatrix
}
\description{
Raw per-protein embeddings generated by UniProt using the ProtT5 model
(prottrans_t5_xl_u50) for the reference proteome of \emph{Escherichia coli}.
}
\details{
The embedding matrix for several species can be downloaded
\href{https://www.uniprot.org/help/downloads#embeddings}{here}. Protein
sequences longer than 12k residues were excluded due to computational
limitations.
}
\examples{
\dontrun{
 eColiEmbeddingMatrix
}
}
\references{
Bateman, A., Martin, M.-J., Orchard, S., Magrane, M., Ahmad, S., Alpi, E.,
Bowler-Barnett, E. H., Britto, R., Bye-A-Jee, H., Cukura, A., Denny, P.,
Dogan, T., Ebenezer, T., Fan, J., Garmiri, P., da Costa Gonzales, L. J.,
Hatton-Ellis, E., Hussein, A., Ignatchenko, A., & Insana, G. (2022).
UniProt: the Universal Protein Knowledgebase in 2023. \emph{Nucleic Acids
Research}, 51(D1). \href{https://doi.org/10.1093/nar/gkac1052}{Link}.
}
\keyword{datasets}
