% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancesRanks.R
\name{getDistancesByMapping}
\alias{getDistancesByMapping}
\title{Get Distances For Mapped Pairs in a Distance Matrix}
\usage{
getDistancesByMapping(distanceMatrix, mapping)
}
\arguments{
\item{distanceMatrix}{A data frame representing the distance matrix from
an embedding matrix. Each value in the distance matrix is the distance
calculated between the row and column embeddings. The distance matrix
should contain protein identifies in the row and columns indices, with
the diagonal all zeroes.}

\item{mapping}{A data frame with two columns of protein identifiers matching
the identifiers used for the embedding matrix.}
}
\value{
Returns a data frame containing the proteins in the matching and
    their associated distances from the distance matrix.
}
\description{
A function that generates a data frame containing the distances for specified
    pairs using a mapping.
}
\examples{
# Generate distance matrix with default setings
eColiDistMatrix <- generateDistanceMatrix(eColiEmbeddingMatrix)

# Retrieve distances by mapping
eColiParalogDistances <- getDistancesByMapping(eColiDistMatrix,
                                                eColiParalogMapping)

}
