% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPairs.R
\name{getFarthestPairs}
\alias{getFarthestPairs}
\title{Get Farthest Pairs in a Rank Matrix}
\usage{
getFarthestPairs(rankMatrix)
}
\arguments{
\item{rankMatrix}{A data frame representing the rank matrix calculated from
an embedding matrix. Each value in the rank matrix is the rank of the
distance of the column protein compared to all other distances computed
with the row protein. These are not reciprocal values, and the diagonal
(where the row and column protein are the same) is always rank 0. The
rank matrix contains protein identifies in the row and columns indices.}
}
\value{
Returns a data frame containing the protein with the farthest rank
    for each protein.
}
\description{
A function that generates a data frame containing the farthest pairs for each
    protein in a ranked matrix.
}
\examples{
# Generate distance matrix with default setings
eColiDistMatrix <- generateDistanceMatrix(eColiEmbeddingMatrix)

# Generate rank matrix from distance matrix
eColiRankMatrix <- generateRankMatrix(eColiDistMatrix)

# Generate farthest pairs matrix
eColiFarthestPairsMatrix <- getFarthestPairs(eColiRankMatrix)

}
