% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEmbeddings.R
\name{loadEmbeddings}
\alias{loadEmbeddings}
\title{Loads Embedding Matrices From Different File Types}
\usage{
loadEmbeddings(path, fileType)
}
\arguments{
\item{path}{A string containing the path to the embedding matrix generated
from a protein Language Model from the root folder. The embedding matrix
must have proteins in the rows with the identifier in the index and
embedding dimensions in the columns. If stored in an h5 file, this
function assumes each row is a dataset under the root path.}

\item{fileType}{A string specifying the file type of the embedding matrix,
must be either "csv", "tsv", or "h5".}
}
\value{
Returns a data frame representing the embedding matrix with proteins
    in the rows and embedding dimensions in the columns.
}
\description{
A function that loads an embedding matrix generated using a protein Language
    Model given a path from the root folder and file type that is either
    "csv", "tsv", or "h5".
}
\examples{
\dontrun{
 loadEmbeddings("path/to/file.csv", "csv")
 loadEmbeddings("path/to/file.tsv", "tsv")
 loadEmbeddings("path/to/file.h5", "h5")
}

}
