% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process Embedding Matrix}
\usage{
processData(embeddingMatrix)
}
\arguments{
\item{embeddingMatrix}{A data frame containing an embedding matrix with
identifiers in the index and columns representing embedding dimensions.}
}
\value{
Returns a data frame representing the embedding matrix with no NULL
    or duplicate values. Protein identifiers are in the rows and embedding
    dimensions in the columns.
}
\description{
A function that removes NULL and duplicate values from a data frame
representing an embedding matrix.
}
\examples{
# Check the original matrix dimensions
embeddingMatrix <- eColiEmbeddingMatrix
print(dim(embeddingMatrix))

# Add NULL values to 100 rows
set.seed(42)
indices_to_nullify <- sample(nrow(embeddingMatrix), 100)
embeddingMatrix[indices_to_nullify, ] <- NA

# Process the data
processedEmbeddingMatrix <- processData(embeddingMatrix)
print(dim(processedEmbeddingMatrix))
}
