% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SARSCoV2Mapping}
\alias{SARSCoV2Mapping}
\title{Example Mapping for \emph{Severe Acute Respiratory Syndrome
Coronavirus 2 (SARS CoV-2)}}
\format{
A data frame with 87 rows and 2 total columns. Columns Protein 1
    and Protein 2 are UniProt identifiers, and each row represents one
    pair of proteins with Uniprot Ids starting with the same 5 characters.
}
\usage{
SARSCoV2Mapping
}
\description{
A one-to-many mapping for \emph{SARS CoV-2} generated for illustrating how to
use the functions in this package.
}
\details{
Mapping was generated by pairing proteins that have the same first 5
characters in their Uniprot Id. This mapping is strictly used for example
purposes.
}
\examples{
\dontrun{
 SARSCoV2Mapping
}
}
\keyword{datasets}
