% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEmbeddings.R
\name{visualizeRankDistribution}
\alias{visualizeRankDistribution}
\title{Plot the Distribution of Embedding Ranks in a Mapping}
\usage{
visualizeRankDistribution(rankMatrix, mapping)
}
\arguments{
\item{rankMatrix}{A data frame representing the rank matrix calculated from
a distance matrix. The distance matrix is a data frame computed from
an embedding matrix that has been generated from a protein Language
Model. Each value in the rank matrix is the rank of the
distance of the column protein compared to all other distances computed
with the row protein. These are not reciprocal values, and the diagonal
(where the row and column protein are the same) is always rank 0. The
rank matrix contains protein identifies in the row and columns indices.}

\item{mapping}{A data frame with two columns of protein identifiers matching
the identifiers used for the embedding matrix.}
}
\value{
Produces a histogram of the distribution of embedding ranks in a
    mapping.
}
\description{
A function that plots the distribution of embedding ranks from a rank matrix
    according to a mapping.
}
\examples{
# Generate distance matrix with default setings
distMatrix <- generateDistanceMatrix(SARSCoV2EmbeddingMatrix)

# Generate rank matrix from distance matrix
rankMatrix <- generateRankMatrix(distMatrix)

# Visualize rank distribution by mapping
mapping <- SARSCoV2Mapping
visualizeRankDistribution(rankMatrix, mapping)

}
