% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eColiEmbeddingMatrix}
\alias{eColiEmbeddingMatrix}
\title{Raw Per-Protein Embeddings For Escherichia coli}
\format{
A data frame with 4,401 rows and 1024 total columns.
\describe{
 \item{Index}{UniprotKB identifier for each protein.}
 \item{Columns}{Each embedding dimension from 1 to 1024.}
}
}
\source{
UniProt.

https://www.uniprot.org/help/downloads#embeddings
}
\usage{
eColiEmbeddingMatrix
}
\description{
Raw per-protein embeddings generated by UniProt using the ProtT5 model
(prottrans_t5_xl_u50) for the reference proteome of Escherichia coli. Protein
sequences longer than 12k residues were excluded due to computational
limitations.
}
\examples{
\dontrun{
 eColiEmbeddingMatrix
}
}
\keyword{datasets}
