% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eColiEmbeddingMatrix}
\alias{eColiEmbeddingMatrix}
\title{Raw Per-Protein Embeddings For Escherichia coli}
\format{
A data frame with 4,401 rows and 1024 total columns. In the index are
    UniProt identifiers for each protein, and columns are labelled V1, V2,
    ..., V1024 to represent each dimension of the embedding.
}
\source{
UniProt.

https://www.uniprot.org/help/downloads#embeddings
}
\usage{
eColiEmbeddingMatrix
}
\description{
Raw per-protein embeddings generated by UniProt using the ProtT5 model
(prottrans_t5_xl_u50) for the reference proteome of Escherichia coli. Protein
sequences longer than 12k residues were excluded due to computational
limitations.
}
\examples{
\dontrun{
 eColiEmbeddingMatrix
}
}
\keyword{datasets}
