% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDistanceRankMatrix.R
\name{generateRankMatrix}
\alias{generateRankMatrix}
\title{Generate Rank Matrix}
\usage{
generateRankMatrix(distanceMatrix)
}
\arguments{
\item{distanceMatrix}{A data frame representing the distance matrix from
an embedding matrix. Each value in the distance matrix is the distance
calculated between the row and column embeddings. The distance matrix
should contain protein identifies in the row and columns indices, with
the diagonal all zeroes.}
}
\value{
Returns a data frame representing the rank matrix calculated from
    the distance matrix. Each value in the rank matrix is the rank of the
    distance of the column protein compared to all other distances computed
    with the row protein. These are not reciprocal values, and the diagonal
    (where the row and column protein are the same) is always rank 0. The
    rank matrix contains protein identifies in the row and columns indices.
}
\description{
A function that compute a rank matrix from a distance matrix.
}
\examples{
\dontrun{
# Generate distance matrix with default setings
embeddingMatrix <- eColiEmbeddingMatrix
eColiDistMatrix <- generateDistanceMatrix(embeddingMatrix)

# Generate rank matrix from distance matrix
eColiRankMatrix <- generateRankMatrix(eColiDistMatrix)
}
}
