% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancesRanks.R
\name{getDistancesByMapping}
\alias{getDistancesByMapping}
\title{Get Distances For Mapped Pairs in a Distance Matrix}
\usage{
getDistancesByMapping(distanceMatrix, mapping)
}
\arguments{
\item{distanceMatrix}{A data frame representing the distance matrix from
an embedding matrix that has been generated from a protein Language
Model. Each value in the distance matrix is the distance calculated
between the row and column embeddings. The distance matrix should contain
protein identifies in the row and columns indices, with the diagonal all
zeroes.}

\item{mapping}{A data frame with two columns of protein identifiers matching
the identifiers used for the embedding matrix.}
}
\value{
Returns a data frame containing the proteins in the matching and
    their associated distances from the distance matrix.
}
\description{
A function that generates a data frame containing the distances from a
    distance matrix between pairs from a user-specified mapping.
}
\examples{
embeddingMatrix <- SARSCoV2EmbeddingMatrix
distMatrix <- generateDistanceMatrix(embeddingMatrix)

# Retrieve distances by mapping
mapping <- SARSCoV2Mapping
distances <- getDistancesByMapping(distMatrix, mapping)

}
