% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistancesRanks.R
\name{getRanksByMapping}
\alias{getRanksByMapping}
\title{Get Ranks For Mapped Pairs in a Rank Matrix}
\usage{
getRanksByMapping(rankMatrix, mapping)
}
\arguments{
\item{rankMatrix}{A data frame representing the rank matrix calculated from
a distance matrix. The distance matrix is a data frame computed from
an embedding matrix that has been generated from a protein Language
Model. Each value in the rank matrix is the rank of the
distance of the column protein compared to all other distances computed
with the row protein. These are not reciprocal values, and the diagonal
(where the row and column protein are the same) is always rank 0. The
rank matrix contains protein identifies in the row and columns indices.}

\item{mapping}{A data frame with two columns of protein identifiers matching
the identifiers used for the embedding matrix.}
}
\value{
Returns a data frame containing the proteins in the matching and
    their associated rank from the rank matrix.
}
\description{
A function that generates a data frame containing the ranks from a rank
    matrix between pairs from a user-specified mapping.
}
\examples{
# Generate distance matrix with default setings
embeddingMatrix <- SARSCoV2EmbeddingMatrix
distMatrix <- generateDistanceMatrix(embeddingMatrix)

# Generate rank matrix from distance matrix
rankMatrix <- generateRankMatrix(distMatrix)

# Retrieve ranks by mapping
mapping <- SARSCoV2Mapping
ranks <- getRanksByMapping(rankMatrix, mapping)

}
