% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process Embedding Matrix}
\usage{
processData(embeddingMatrix)
}
\arguments{
\item{embeddingMatrix}{A data frame containing an embedding matrix generated
from a protein Lanugage Model with identifiers in the index and columns
representing embedding dimensions.}
}
\value{
Returns a data frame representing the embedding matrix with no NULL
    or duplicate values. Protein identifiers are in the rows and embedding
    dimensions in the columns.
}
\description{
A function that removes NULL and duplicate values from a data frame
representing an embedding matrix generated from a protein Language Model.
}
\examples{
# Process the embedding matrix
embeddingMatrix <- SARSCoV2EmbeddingMatrix
processedEmbeddingMatrix <- processData(embeddingMatrix)

}
\references{
Müller K, Wickham H (2023). tibble: Simple Data Frames. R package version
3.2.1.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023). dplyr: A Grammar
of Data Manipulation. R package version 1.1.3.
}
