% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProtemalyze.R
\name{runProtemalyze}
\alias{runProtemalyze}
\title{Launch Shiny App for protemalyze}
\usage{
runProtemalyze()
}
\value{
None. Opens a Shiny application.
}
\description{
A function that launches the Shiny app for protemalyze.
The app provides an interactive interface for visualizing and analyzing
protein embeddings.
}
\examples{
\dontrun{

protemalyze::runProtemalyze()
}

}
