% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runProtemalyze.R
\name{runProtemalyze}
\alias{runProtemalyze}
\title{Launch Shiny App for protemalyze}
\usage{
runProtemalyze()
}
\value{
None. Opens a Shiny webpage
}
\description{
A function that launches the Shiny app for protemalyze.
The purpose of this app is only to illustrate how a Shiny
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

protemalyze::runProtemalyze()
}

}
