% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEmbeddings.R
\name{visualizeDistanceDistribution}
\alias{visualizeDistanceDistribution}
\title{Plot the Distribution of Embedding Distances in a Mapping}
\usage{
visualizeDistanceDistribution(distanceMatrix, mapping)
}
\arguments{
\item{distanceMatrix}{A data frame representing the distance matrix from
an embedding matrix that has been generated from a protein Language
Model. Each value in the distance matrix is the distance calculated
between the row and column embeddings. The distance matrix should contain
protein identifies in the row and columns indices, with the diagonal all
zeroes.}

\item{mapping}{A data frame with two columns of protein identifiers matching
the identifiers used for the embedding matrix.}
}
\value{
Produces a histogram of the distribution of embedding distances in a
    mapping.
}
\description{
A function that plots the distribution of embedding distances from a distance
    matrix according to a mapping.
}
\examples{
# Generate distance matrix with default setings
distMatrix <- generateDistanceMatrix(SARSCoV2EmbeddingMatrix)

# Visualize distance distribution by mapping
mapping <- SARSCoV2Mapping
visualizeDistanceDistribution(distMatrix, mapping)

}
\references{
Wickham H, Chang W, Henry L, Pedersen TL, Takahashi K, Wilke C, Woo K, Yutani
H, Dunnington D (2023). ggplot2: Create Elegant Data Visualisations Using the
Grammar of Graphics. R package version 3.4.3.
}
