% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEmbeddings.R
\name{visualizeEmbeddingUMAP}
\alias{visualizeEmbeddingUMAP}
\title{Plot an Interactive UMAP From an Embedding Matrix}
\usage{
visualizeEmbeddingUMAP(embeddingMatrix)
}
\arguments{
\item{embeddingMatrix}{A data frame containing an embedding matrix with
identifiers in the index and columns representing embedding dimensions.
It is recommended that you utilize protemalyze::processData to clean the
embedding matrix.}
}
\value{
Produces an interactive plot of the UMAP created from an
    embedding matrix.
}
\description{
A function that creates an interactive plot of the UMAP created from an
    embedding matrix generated from a protein Language Model.
}
\examples{
# Plot an interactive UMAP
embeddingMatrix <- SARSCoV2EmbeddingMatrix
visualizeEmbeddingUMAP(embeddingMatrix)

}
\references{
Sievert C (2023). plotly: Create Interactive Web Graphics via 'plotly.js'. R
package version 4.10.2.

Konopka T (2023). umap: Uniform Manifold Approximation and Projection. R
package version 0.2.10.0
}
