% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEmbeddings.R
\name{visualizeEmbeddingUMAP}
\alias{visualizeEmbeddingUMAP}
\title{Plot an Interactive UMAP From an Embedding Matrix}
\usage{
visualizeEmbeddingUMAP(embeddingMatrix)
}
\arguments{
\item{embeddingMatrix}{A data frame containing an embedding matrix with
identifiers in the index and columns representing embedding dimensions.}
}
\value{
Produces an interactive plot of the UMAP created from an
    embedding matrix.
}
\description{
A function that creates an interactive plot of the UMAP created from an
    embedding matrix.
}
\examples{
\dontrun{
# Plot an interactive UMAP
embeddingMatrix <- eColiEmbeddingMatrix
visualizeEmbeddingUMAP(embeddingMatrix)
}

}
